:recycle: refactor : Multipart 파일 데이터를 Fixture로 분리하고 file 업로드 API 직접 호출하도록 변경
