+++ apps/localtalk-api/src/test/kotlin/com/localtalk/api/file/entrypoint/FileControllerTest.kt
@@ -9,9 +9,6 @@ import org.junit.jupiter.api.Test
 import org.springframework.core.io.ByteArrayResource
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
-import org.springframework.http.client.MultipartBodyBuilder
-import org.springframework.mock.web.MockMultipartFile
-import org.springframework.test.web.reactive.server.WebTestClient
 import org.springframework.util.LinkedMultiValueMap
 import org.springframework.web.reactive.function.BodyInserters
 
@@ -25,8 +22,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
 
             val testFile = FileFixture.createMockFile("test.jpg", "image/jpeg")
+            val multipartData = FileFixture.createMultipartData(testFile, FileType.PROFILE.name)
 
-            val response = uploadFile(testFile, FileType.PROFILE.name)
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isOk
@@ -44,8 +46,13 @@ class FileControllerTest : IntegrationTest() {
         fun `PNG 이미지 파일이면 파일 URL을 반환한다`() {
             loginAsTemporaryMember()
             val testFile = FileFixture.createMockFile("test.png", "image/png")
+            val multipartData = FileFixture.createMultipartData(testFile, "PROFILE")
 
-            val response = uploadFile(testFile, "PROFILE")
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isOk
@@ -69,7 +76,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
             val testFile = FileFixture.createMockFile("test.jpg", "image/jpeg")
 
-            val response = uploadFile(testFile, invalidType)
+            val multipartData = FileFixture.createMultipartData(testFile, invalidType)
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isBadRequest
@@ -83,7 +96,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
             val testFile = FileFixture.createMockFile("test.pdf", "application/pdf")
 
-            val response = uploadFile(testFile, "PROFILE")
+            val multipartData = FileFixture.createMultipartData(testFile, "PROFILE")
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isBadRequest
@@ -97,7 +116,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
             val testFile = FileFixture.createMockFile("test.txt", "text/plain")
 
-            val response = uploadFile(testFile, "PROFILE")
+            val multipartData = FileFixture.createMultipartData(testFile, "PROFILE")
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isBadRequest
@@ -115,7 +140,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
             val emptyFile = FileFixture.createMockFile("empty-file.jpg", "image/jpeg", size = 0)
 
-            val response = uploadFile(emptyFile, "PROFILE")
+            val multipartData = FileFixture.createMultipartData(emptyFile, "PROFILE")
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isBadRequest
@@ -129,7 +160,13 @@ class FileControllerTest : IntegrationTest() {
             loginAsTemporaryMember()
             val largeTestFile = FileFixture.createMockFile("large-test-file.jpg", "image/jpeg", size = 11L * 1024 * 1024)
 
-            val response = uploadFile(largeTestFile, "PROFILE")
+            val multipartData = FileFixture.createMultipartData(largeTestFile, "PROFILE")
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().is4xxClientError
@@ -187,26 +224,16 @@ class FileControllerTest : IntegrationTest() {
         fun `인증되지 않은 사용자면 401 에러를 반환한다`() {
             val testFile = FileFixture.createMockFile("test.jpg", "image/jpeg")
 
-            val response = uploadFile(testFile, "PROFILE")
+            val multipartData = FileFixture.createMultipartData(testFile, "PROFILE")
+
+            val response = webTestClient.post()
+                .uri("/api/v1/files")
+                .contentType(MediaType.MULTIPART_FORM_DATA)
+                .body(BodyInserters.fromMultipartData(multipartData))
+                .exchange()
 
             response
                 .expectStatus().isUnauthorized
         }
-    }
-    private fun uploadFile(file: MockMultipartFile, type: String): WebTestClient.ResponseSpec {
-        val bodyBuilder = MultipartBodyBuilder()
-
-        bodyBuilder.part("file", file.resource)
-            .filename(file.originalFilename)
-            .contentType(MediaType.parseMediaType(file.contentType!!))
-
-        bodyBuilder.part("type", type)
-            .contentType(MediaType.TEXT_PLAIN)
-
-        return webTestClient.post()
-            .uri("/api/v1/files")
-            .contentType(MediaType.MULTIPART_FORM_DATA)
-            .body(BodyInserters.fromMultipartData(bodyBuilder.build()))
-            .exchange()
     }
 }
