+++ apps/localtalk-api/src/test/kotlin/com/localtalk/api/support/fixture/FileFixture.kt
@@ -1,6 +1,10 @@
 package com.localtalk.api.support.fixture
 
+import org.springframework.http.HttpEntity
+import org.springframework.http.MediaType
+import org.springframework.http.client.MultipartBodyBuilder
 import org.springframework.mock.web.MockMultipartFile
+import org.springframework.util.MultiValueMap
 
 object FileFixture {
 
@@ -14,4 +18,20 @@ object FileFixture {
         contentType,
         ByteArray(size.toInt()),
     )
+
+    fun createMultipartData(
+        file: MockMultipartFile,
+        type: String
+    ): MultiValueMap<String, HttpEntity<*>> {
+        val bodyBuilder = MultipartBodyBuilder()
+
+        bodyBuilder.part("file", file.resource)
+            .filename(file.originalFilename)
+            .contentType(MediaType.parseMediaType(file.contentType!!))
+
+        bodyBuilder.part("type", type)
+            .contentType(MediaType.TEXT_PLAIN)
+
+        return bodyBuilder.build()
+    }
 }
